+++ Changelog.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for the visitor pattern,
+  by adding `protocol Visitor`/`protocol Visitable` & `enum VisitorContinueKind`.
 - Added a changelog.
   #17 by @mattt.